Extract ViewComponent::Form::Renderer
